Add qps observability, fractional rates (#125)
- Migrates to using time.perf_counter for consistency across processes
and threads.
- Adds several load report fields to identify load accuracy.
- Changes load config to support fractional rates #123

+++ inference_perf/config.py
@@ -79,7 +79,7 @@ class MetricsClientType(Enum):
 
 
 class LoadStage(BaseModel):
-    rate: int
+    rate: float
     duration: int
